models reorganized
