+++ srimulatorcore/models/__init__.py
@@ -16,5 +16,7 @@
 # 
 # Email mccrea.engineering@gmail.com for questions, comments, or to report bugs.
 
-
-
+from .map import *
+from .robot import *
+from .supervisor import *
+from .world import *
