+++ srimulatorcore/models/map/line_segment.py
@@ -17,8 +17,8 @@
 # Email mccrea.engineering@gmail.com for questions, comments, or to report bugs.
 
 
-from ..utils import linalg2_util as linalg
-from .geometry import Geometry
+from ...utils import linalg2_util as linalg
+from ...utils.geometry import Geometry
 
 class LineSegment( Geometry ):
