+++ srimulatorcore/models/robot/controllers/avoid_obstacles_controller.py
