+++ srimulatorcore/models/robot/controllers/go_to_goal_controller.py
