+++ srimulatorcore/models/robot/robot.py
@@ -16,7 +16,7 @@
 # 
 # Email mccrea.engineering@gmail.com for questions, comments, or to report bugs.
 
-from .pose import Pose
+from ...utils.pose import Pose
 
 class Robot:
