+++ srimulatorcore/sobots_samples/kheperaiii/controllers/gtg_and_ao_controller.py
@@ -19,7 +19,7 @@
 
 from math import atan2, pi
 
-from ....models.controllers.gtg_and_ao_controller import GTGAndAOController
+from ....models.robot.controllers.gtg_and_ao_controller import GTGAndAOController
 from ....utils import linalg2_util as linalg
 from .avoid_obstacles_controller import KheperaiiiAvoidObstaclesController
 from .go_to_goal_controller import KheperaiiiGoToGoalController
