+++ srimulatorcore/utils/geometry.py
